requirements.txt
django
django-filter
djangorestframework
djangorestframework-jwt